Separate fs and memmap features
On some targets such as wasm32-wasi, memmap2 is not supported but fs
works fine. This commit separates memmap use from the fs feature into a
new memmap feature and adds a fallback for non-memmap fs builds.

This allows fontdb to be used with fs support on wasm32-wasi targets.