Simplify fontdb::Source enum
Remove the recently added SharedBinary variant again and instead store
an Arc<dyn AsRef<[u8]>> for the Binary variant. This also allows
replacing the Arc<Source> with just Source in the FaceInfo, as Source
can now be cloned without cloning the data.

The make_shared_face_data and make_face_data_unshared functions are now
feature gated to memmap as they don't make any sense otherwise anymore.
Although while a Source::File could be turned into a Source::Binary, it
would not be possible to unshare them again because the path is not
stored.